chore: consistent searchbar across folders and lists
