+++ packages/ui/src/elements/ListControls/index.scss
@@ -6,99 +6,19 @@
     flex-direction: column;
     gap: 2px;
 
-    &__wrap {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: base(0.5);
-      background-color: var(--theme-elevation-50);
-      border-radius: var(--style-radius-m);
-    }
-
-    &__search {
-      display: flex;
-      background-color: var(--theme-elevation-50);
-      border-radius: var(--style-radius-m);
-      gap: base(0.4);
-      flex-grow: 1;
-      position: relative;
-
-      .icon {
-        flex-shrink: 0;
-      }
-    }
-
-    .icon--search {
-      position: absolute;
-      left: 0;
-      top: 50%;
-      transform: translate3d(0, -50%, 0);
-      inset-inline-start: base(0.4);
-      z-index: 1;
-      pointer-events: none;
-    }
-
-    .search-filter {
-      flex-grow: 1;
-
-      input {
-        height: 46px;
-        padding-left: 36px;
-        margin: 0;
-      }
-    }
-
-    &__custom-control {
-      padding: 0;
-      border-radius: 0;
-    }
-
-    &__buttons {
-      display: flex;
-      align-items: center;
-      gap: calc(var(--base) / 4);
-      padding-right: 10px;
-    }
-
     .pill-selector,
     .where-builder,
     .sort-complex,
     .group-by-builder {
-      margin-top: base(1);
+      margin-top: calc(var(--base) / 2);
     }
 
     @include small-break {
-      &__wrap {
-        flex-direction: column;
-        align-items: stretch;
-        background-color: transparent;
-        border-radius: 0;
-      }
-
-      .search-filter {
-        width: 100%;
-
-        input {
-          height: 40px;
-          padding: 0 base(1.5);
-        }
-      }
-
-      &__buttons {
-        padding-right: 0;
-        margin: 0;
-        width: 100%;
-
+      .search-bar__actions {
         .pill {
           padding: base(0.2) base(0.2) base(0.2) base(0.4);
           justify-content: space-between;
         }
-      }
-
-      .pill-selector,
-      .where-builder,
-      .sort-complex {
-        margin-top: calc(var(--base) / 2);
       }
 
       &__toggle-columns,
